feat(gui): start polishing gui
Added a play button that opens player-form dialog. The play button is
removed and the form dialog is closed when the form is confirmed.
The grid and it's cells are dynamically built when the game starts and
can be reset upon clicking the reset button.
Removed ability to play via the console as the game can be played purely
via the gui, although winner is still only displayed in the console.
-- Bugs to fix --
Ongoing bug: winner *sometimes* isn't determined until following round.
New bug: form dialog cancel button isn't working.